split parsers into auto-discovered modules
