+++ _version.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-__version__ = "0.4.5"
+__version__ = "0.5.0"
 __package__ = 'metadata-flatten'
 __description__ = "Metadata Flattening Service"
 __copyright__ = "Copyright AT&T Services and Warner Media 2020"
